Move `DomainSettingsViewModel` to a separate file. Add comments.
