+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/Domains/DomainSettingsView.swift
@@ -2,8 +2,6 @@ import SwiftUI
 
 /// Hosting controller that wraps the `DomainSettingsView` view.
 final class DomainSettingsHostingController: UIHostingController<DomainSettingsView> {
-    /// - Parameters:
-    ///   - viewModel: View model for the domain settings.
     init(viewModel: DomainSettingsViewModel) {
         super.init(rootView: DomainSettingsView(viewModel: viewModel))
     }
@@ -19,73 +17,7 @@ final class DomainSettingsHostingController: UIHostingController<DomainSettingsV
     }
 }
 
-final class DomainSettingsViewModel: ObservableObject {
-    struct Domain {
-        /// Whether the domain is the site's primary domain.
-        let isPrimary: Bool
-
-        /// The address of the domain.
-        let name: String
-
-        // The next renewal date.
-        let autoRenewalDate: Date?
-    }
-
-    struct FreeStagingDomain {
-        /// Whether the domain is the site's primary domain.
-        let isPrimary: Bool
-
-        /// The address of the domain.
-        let name: String
-    }
-
-    @Published private(set) var hasDomainCredit: Bool = false
-    @Published private(set) var domains: [Domain] = []
-    @Published private(set) var freeStagingDomain: FreeStagingDomain?
-
-    private let siteID: Int64
-    private let stores: StoresManager
-
-    init(siteID: Int64, stores: StoresManager = ServiceLocator.stores) {
-        self.siteID = siteID
-        self.stores = stores
-    }
-
-    func onAppear() {
-        stores.dispatch(DomainAction.loadDomains(siteID: siteID) { [weak self] result in
-            self?.handleDomainsResult(result)
-        })
-
-        stores.dispatch(PaymentAction.loadSiteCurrentPlan(siteID: siteID) { [weak self] result in
-            self?.handleSiteCurrentPlanResult(result)
-        })
-    }
-}
-
-private extension DomainSettingsViewModel {
-    func handleDomainsResult(_ result: Result<[SiteDomain], Error>) {
-        switch result {
-        case .success(let domains):
-            let stagingDomain = domains.first(where: { $0.renewalDate == nil })
-            freeStagingDomain = stagingDomain
-                .map { FreeStagingDomain(isPrimary: $0.isPrimary, name: $0.name) }
-            self.domains = domains.filter { $0 != stagingDomain }
-                .map { Domain(isPrimary: $0.isPrimary, name: $0.name, autoRenewalDate: $0.renewalDate) }
-        case .failure(let error):
-            DDLogError("⛔️ Error retrieving domains for siteID \(siteID): \(error)")
-        }
-    }
-
-    func handleSiteCurrentPlanResult(_ result: Result<WPComSitePlan, Error>) {
-        switch result {
-        case .success(let sitePlan):
-            hasDomainCredit = sitePlan.hasDomainCredit
-        case .failure(let error):
-            DDLogError("⛔️ Error retrieving site plan for siteID \(siteID): \(error)")
-        }
-    }
-}
-
+/// Shows a site's domains with actions to add a domain or redeem a domain credit.
 struct DomainSettingsView: View {
     @ObservedObject private var viewModel: DomainSettingsViewModel
 
@@ -158,7 +90,7 @@ private extension DomainSettingsView {
 import Yosemite
 import enum Networking.DotcomError
 
-/// StoresManager that specifically handles `DomainAction` for `DomainSettingsView` previews.
+/// StoresManager that specifically handles actions for `DomainSettingsView` previews.
 final class DomainSettingsViewStores: DefaultStoresManager {
     private let domainsResult: Result<[SiteDomain], Error>
     private let sitePlanResult: Result<WPComSitePlan, Error>
